Proof of concept task directory
