+++ allegations/connection.py
